+++ src/lib/stores.js
@@ -1,0 +1,5 @@
+import { writable } from 'svelte/store';
+
+export const accessControl = writable(0);
+
+export const code = writable('');